+++ Trace/Hpc/Tix.hs
@@ -1,11 +1,4 @@
-{-# LANGUAGE CPP #-}
-#if __GLASGOW_HASKELL__ >= 704
-{-# LANGUAGE Safe #-}
-#elif __GLASGOW_HASKELL__ >= 702
--- System.FilePath in filepath version 1.2.0.1 isn't marked or implied Safe,
--- as shipped with GHC 7.2.
-{-# LANGUAGE Trustworthy #-}
-#endif
+{-# LANGUAGE Safe, DeriveGeneric, StandaloneDeriving #-}
 ------------------------------------------------------------
 -- Andy Gill and Colin Runciman, June 2006
 ------------------------------------------------------------
@@ -16,6 +9,8 @@ module Trace.Hpc.Tix(Tix(..), TixModule(..),
                      tixModuleName, tixModuleHash, tixModuleTixs, tixModuleTraceInfo, tixModuleTrace,
                      readTix, writeTix, getTixFileName) where
 
+import Control.DeepSeq (NFData)
+import GHC.Generics (Generic)
 import System.FilePath (replaceExtension)
 
 import Trace.Hpc.Util (Hash, catchIO, readFileUtf8, writeFileUtf8)
@@ -24,6 +19,11 @@ import Trace.Hpc.Util (Hash, catchIO, readFileUtf8, writeFileUtf8)
 -- what boxes are ticked.
 data Tix = Tix [TixModule]
         deriving (Read, Show, Eq)
+
+-- | @since 0.6.2.0
+deriving instance (Generic Tix)
+-- | @since 0.6.2.0
+instance NFData Tix
 
 data TixModule = TixModule
                  String    --  module name
@@ -34,6 +34,11 @@ data TixModule = TixModule
                  [Integer] --  current trace posistion
                  [Integer] -- traces
         deriving (Read, Show, Eq)
+
+-- | @since 0.6.2.0
+deriving instance (Generic TixModule)
+-- | @since 0.6.2.0
+instance NFData TixModule
 
 -- TODO: Turn extractors below into proper 'TixModule' field-labels
 tixModuleName :: TixModule -> String
@@ -50,19 +55,23 @@ tixModuleTrace (TixModule  _ _ _ _ _ trace) = trace
 -- We /always/ read and write Tix from the current working directory.
 
 -- | Read a @.tix@ File.
-readTix :: String
+readTix :: FilePath
         -> IO (Maybe Tix)
 readTix tixFilename =
   catchIO (fmap (Just . read) $ readFileUtf8 tixFilename)
           (const $ return Nothing)
 
 -- | Write a @.tix@ File.
-writeTix :: String
+writeTix :: FilePath
          -> Tix
          -> IO ()
 writeTix name tix = writeFileUtf8 name (show tix)
 
 -- | 'getTixFullName' takes a binary or @.tix@-file name,
 -- and normalizes it into a @.tix@-file name.
-getTixFileName :: String -> String
+--
+-- > getTixFileName "example.hs" == "example.tix"
+-- > getTixFileName "example.tar.gz" == "example.tar.tix"
+-- > getTixFileName "example.tix" == "example.tix"
+getTixFileName :: FilePath -> FilePath
 getTixFileName str = replaceExtension str "tix"