+++ changelog.md
@@ -1,8 +1,15 @@
 # Changelog for [`hpc` package](http://hackage.haskell.org/package/hpc)
 
-## 0.7.0.0  *May 2022*
 
+## Unreleased
+
+  * Replace uses of `String` by `FilePath` in functions `readTix`, `writeTix`, `getTixFileName`, `readMix` and `mixCreate`.
+  * Remove support for version of GHC below 8.6
   * Update to work with new hpc based traces
+
+## 0.6.2.0  *September 2021*
+
+  * Addition of `NFData` instances for `Tix` and `TixModule`
 
 ## 0.6.1.0  *October 2019*